+++ CHANGELOG.md
@@ -1,15 +1,20 @@
-### 2.0.1 (2017-11-10)
+### v2.1.0 (2017-11-10)
+
+  * Added + tested support for ESP32
+  * Improved display messages
+
+### v2.0.1 (2017-11-10)
 
   * Added installation instructions to README.md
   * Added problem solving instructions
 
-### 2.0.0 (2017-11-10)
+### v2.0.0 (2017-11-10)
 
   * Renamed from JV_OTA to JeVe_EasyOTA
   * Improved OTA_U8g2 example
   * Added README and CHANGELOG
 
-### 1.0.0 (2017-11-18)
+### v1.0.0 (2017-11-18)
 
   * First working version
   * Added U8g2 example